move EnsureOneOf, EnsureBothOf to the frame_support::traits
